fix: permissions bit is int64. not int

+++ discord/resource_discord_channel_permission.go
@@ -123,8 +123,8 @@ func resourceChannelPermissionUpdate(ctx context.Context, d *schema.ResourceData
 
 	if err := client.ChannelPermissionSet(
 		channelId, overwriteId, permissionType,
-		d.Get("allow").(int64),
-		d.Get("deny").(int64), discordgo.WithContext(ctx)); err != nil {
+		int64(d.Get("allow").(int)),
+		int64(d.Get("deny").(int)), discordgo.WithContext(ctx)); err != nil {
 		return diag.Errorf("Failed to update channel permissions %s: %s", channelId, err.Error())
 	} else {
 		d.SetId(strconv.Itoa(
